+++ CHANGELOG.md
@@ -4,13 +4,16 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - Unit prefixes and operator overloads for electrical units
+- Use GitHub as source server (#12)
 
 ### Changed
+- Rename DegreeKelvin to Kelvin and delete "degK" unit (#30)
 - Case sensitive unit prefixes
 - Switch PCL project to profile 328
+- Refactor UnitProvider.RegisterUnits to extension methods (#6)
 
 ### Removed
--
+- 
 
 ### Fixed
-- 
+- Support parsing of "'" units (#20)